unified the "ZERO" rule normalizing the single CPU/RAM/DISK ranks to the total

+++ src/main/java/org/indigo/cloudproviderruleengine/RankedCloudProvider.java
@@ -1,11 +1,11 @@
 package org.indigo.cloudproviderruleengine;
 
 public class RankedCloudProvider {
-	private int id;
-	private String name;
-	private long rank;
+	private int     id;
+	private String  name;
+	private double  rank;
 	
-	public RankedCloudProvider( int id, String name, long rank) {
+	public RankedCloudProvider( int id, String name, double rank) {
 		this.id   = id;
 		this.name = name;
 		this.rank = rank;
